multi_agent_job_crawler/requirements.txt
firecrawl-py
openai
anthropic
python-dotenv
pandas
numpy
scikit-learn
matplotlib